Fix: Id type from integer to long to match database schema and resolve related errors

+++ src/main/java/com/example/flower_shop/controller/BouquetController.java
@@ -1,9 +1,7 @@
 package com.example.flower_shop.controller;
 
 import com.example.flower_shop.model.Bouquet;
-import com.example.flower_shop.model.Flower;
 import com.example.flower_shop.repository.BouquetRepository;
-import com.example.flower_shop.repository.FlowerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,7 +22,7 @@ public class BouquetController {
     }
 
     @GetMapping("/{id}")
-    public Bouquet getBouquetById(@PathVariable Integer id){
+    public Bouquet getBouquetById(@PathVariable Long id){
         return bouquetRepository.findById(id).orElse(null);
     }
 
@@ -34,12 +32,12 @@ public class BouquetController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBouquet(@PathVariable Integer id){
+    public void deleteBouquet(@PathVariable Long id){
         bouquetRepository.deleteById(id);
     }
     
     @PutMapping("/{id}")
-    public Bouquet updateBouquet(@PathVariable Integer id, @RequestBody Bouquet bouquetDetails) {
+    public Bouquet updateBouquet(@PathVariable Long id, @RequestBody Bouquet bouquetDetails) {
         Bouquet bouquet = bouquetRepository.findById(id).orElse(null);
         if(bouquet != null){
             bouquet.setEmertimi(bouquetDetails.getEmertimi());
